@@ -30,18 +30,18 @@
   },
   "private": false,
   "dependencies": {
-    "@angular/animations": "~7.1.0",
-    "@angular/common": "~7.1.0",
-    "@angular/compiler": "~7.1.0",
-    "@angular/core": "~7.1.0",
-    "@angular/forms": "~7.1.0",
-    "@angular/http": "~7.1.0",
-    "@angular/platform-browser": "~7.1.0",
-    "@angular/platform-browser-dynamic": "~7.1.0",
-    "@angular/platform-server": "^7.1.4",
-    "@angular/pwa": "^0.11.3",
-    "@angular/router": "~7.1.0",
-    "@angular/service-worker": "~7.1.0",
+    "@angular/animations": "~7.2.0",
+    "@angular/common": "~7.2.0",
+    "@angular/compiler": "~7.2.0",
+    "@angular/core": "~7.2.0",
+    "@angular/forms": "~7.2.0",
+    "@angular/http": "~7.2.0",
+    "@angular/platform-browser": "~7.2.0",
+    "@angular/platform-browser-dynamic": "~7.2.0",
+    "@angular/platform-server": "^7.2.0",
+    "@angular/pwa": "^0.12.0",
+    "@angular/router": "~7.2.0",
+    "@angular/service-worker": "~7.2.0",
     "@ng-bootstrap/ng-bootstrap": "^4.0.1",
     "@ng-select/ng-select": "^2.13.2",
     "@ngrx/effects": "~7.0.0",
@@ -52,18 +52,18 @@
     "@types/googlemaps": "^3.30.13",
     "bootstrap": "^4.1.3",
     "core-js": "^2.5.7",
-    "hamburgers": "^0.9.3",
+    "hamburgers": "^1.1.3",
     "ngrx-store-localstorage": "^5.1.0",
     "rxjs": "^6.3.3",
     "ts-loader": "^5.3.2",
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.11.0",
-    "@angular-devkit/build-ng-packagr": "~0.11.0",
-    "@angular/cli": "~7.1.4",
-    "@angular/compiler-cli": "~7.1.0",
-    "@angular/language-service": "~7.1.0",
+    "@angular-devkit/build-angular": "~0.12.0",
+    "@angular-devkit/build-ng-packagr": "~0.12.0",
+    "@angular/cli": "~7.2.1",
+    "@angular/compiler-cli": "7.2.0",
+    "@angular/language-service": "~7.2.0",
     "@ngrx/store-devtools": "~7.0.0",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
