@@ -4,9 +4,9 @@ import { RouterTestingModule } from '@angular/router/testing';
 import { By } from '@angular/platform-browser';
 import { of } from 'rxjs';
 import { ResponsiveBannerComponent } from './responsive-banner.component';
-import { CmsConfig, TranslateUrlOptions } from '@spartacus/core';
+import { CmsConfig, TranslateUrlOptions, Component } from '@spartacus/core';
 import { GenericLinkComponent } from '../../ui/components/generic-link/generic-link.component';
-import { CmsService } from '@spartacus/core';
+import { CmsComponentData } from '../../cms/components/cms-component-data';
 
 const UseCmsModuleConfig: CmsConfig = {
   cmsComponents: {
@@ -67,8 +67,10 @@ describe('ResponsiveBannerComponent', () => {
     urlLink: '/OpenCatalogue/Cameras/Digital-Cameras/Digital-SLR/c/578'
   };
 
-  const MockCmsService = {
-    getComponentData: () => of(componentData)
+  const MockCmsComponentData = <CmsComponentData<Component>>{
+    data$: of(componentData),
+    uid: 'test',
+    contextParameters: null
   };
 
   beforeEach(async(() => {
@@ -80,7 +82,7 @@ describe('ResponsiveBannerComponent', () => {
         MockTranslateUrlPipe
       ],
       providers: [
-        { provide: CmsService, useValue: MockCmsService },
+        { provide: CmsComponentData, useValue: MockCmsComponentData },
         { provide: CmsConfig, useValue: UseCmsModuleConfig }
       ]
     }).compileComponents();
@@ -97,36 +99,31 @@ describe('ResponsiveBannerComponent', () => {
   });
 
   it('should contain responsive banner image source, source set and redirect url', () => {
-    expect(responsiveBannerComponent.component).toBeNull();
-    responsiveBannerComponent.onCmsComponentInit(
-      componentData.uid,
-      of(componentData)
-    );
-    expect(responsiveBannerComponent.component).toBe(componentData);
+    fixture.detectChanges();
     expect(el.query(By.css('a')).nativeElement.href).toContain(
-      '/translated-path' + responsiveBannerComponent.component.urlLink
+      '/translated-path' + componentData.urlLink
     );
     expect(el.query(By.css('img')).nativeElement.src).toContain(
-      responsiveBannerComponent.component.media.desktop.url
+      componentData.media.desktop.url
     );
     expect(el.query(By.css('img')).nativeElement.srcset).toContain(
-      '/medias/Elec-480x320-HomeSpeed-EN-01-480W.jpg 200w,'
+      `${componentData.media.mobile.url} 200w,`
     );
     expect(el.query(By.css('img')).nativeElement.srcset).toContain(
-      '/medias/Elec-770x350-HomeSpeed-EN-01-770W.jpg 500w,'
+      `${componentData.media.tablet.url} 500w,`
     );
     expect(el.query(By.css('img')).nativeElement.srcset).toContain(
-      '/medias/Elec-960x330-HomeSpeed-EN-01-960W.jpg 800w,'
+      `${componentData.media.desktop.url} 800w,`
     );
     expect(el.query(By.css('img')).nativeElement.srcset).toContain(
-      '/medias/Elec-1400x440-HomeSpeed-EN-01-1400W.jpg 1200w'
+      `${componentData.media.widescreen.url} 1200w`
     );
     expect(el.query(By.css('picture')).nativeElement.classList[0]).toBe(
       'responsive-banner'
     );
 
     expect(el.query(By.css('picture')).nativeElement.classList[1]).toBe(
-      componentData.uid
+      MockCmsComponentData.uid
     );
 
     expect(el.query(By.css('picture')).nativeElement.innerHTML).toContain(
