@@ -3,15 +3,15 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
 import { RouterTestingModule } from '@angular/router/testing';
 
-import { ProductService, Product, CmsComponent } from '@spartacus/core';
+import { ProductService, Product, Component } from '@spartacus/core';
 
 import { of, Observable } from 'rxjs';
 
 import { BootstrapModule } from '../../bootstrap.module';
-import { CmsService } from '@spartacus/core';
 import { PictureComponent } from '../../ui/components/media/picture/picture.component';
 
 import { ProductCarouselComponent } from './product-carousel.component';
+import { CmsComponentData } from '../../cms/components/cms-component-data';
 
 @Pipe({
   name: 'cxTranslateUrl'
@@ -43,11 +43,9 @@ const mockProduct: Product = {
   }
 };
 
-class MockCmsService {
-  getComponentData<T extends CmsComponent>(): Observable<T> {
-    return of(mockComponentData);
-  }
-}
+const MockCmsComponentData = <CmsComponentData<Component>>{
+  data$: of(mockComponentData)
+};
 
 class MockProductService {
   get(): Observable<any> {
@@ -73,7 +71,7 @@ describe('ProductCarouselComponent', () => {
         MockTranslateUrlPipe
       ],
       providers: [
-        { provide: CmsService, useClass: MockCmsService },
+        { provide: CmsComponentData, useValue: MockCmsComponentData },
         { provide: ProductService, useClass: MockProductService }
       ]
     }).compileComponents();
@@ -81,7 +79,6 @@ describe('ProductCarouselComponent', () => {
 
   beforeEach(() => {
     fixture = TestBed.createComponent(ProductCarouselComponent);
-    fixture.detectChanges();
     productCarouselComponent = fixture.componentInstance;
     fixture.detectChanges();
     el = fixture.debugElement;
@@ -91,26 +88,18 @@ describe('ProductCarouselComponent', () => {
     expect(productCarouselComponent).toBeTruthy();
   });
 
-  it('should have empty productCodes', () => {
-    expect(productCarouselComponent.productCodes).toBeFalsy();
-  });
-
   it('should have productCodes', () => {
-    expect(productCarouselComponent.component).toBeNull();
-    productCarouselComponent.onCmsComponentInit(mockComponentData.uid);
     expect(productCarouselComponent.productCodes).toEqual(productCodeArray);
   });
 
   it('should have 1 group', () => {
     spyOn<any>(productCarouselComponent, 'getItemsPerPage').and.returnValue(4);
-    productCarouselComponent.onCmsComponentInit(mockComponentData.uid);
     expect(productCarouselComponent.productGroups.length).toBe(1);
   });
 
   it('should contain cms content in the html rendering after bootstrap', () => {
-    productCarouselComponent.onCmsComponentInit(mockComponentData.uid);
     expect(
       el.query(By.css('.cx-carousel__header')).nativeElement.textContent
-    ).toEqual(productCarouselComponent.component.title);
+    ).toContain(mockComponentData.title);
   });
 });
