@@ -1,22 +1,22 @@
 import {
   Component,
-  Input,
   ChangeDetectionStrategy,
   OnDestroy,
   ViewChild,
-  OnInit,
-  ChangeDetectorRef
+  OnInit
 } from '@angular/core';
 
-import { ProductService, Product, WindowRef } from '@spartacus/core';
+import {
+  ProductService,
+  Product,
+  CmsProductCarouselComponent,
+  WindowRef
+} from '@spartacus/core';
 
 import { Subscription, fromEvent, Observable } from 'rxjs';
 import { debounceTime } from 'rxjs/operators';
-
-import { AbstractCmsComponent } from '../../cms/components/abstract-cms-component';
-
-import { CmsService } from '@spartacus/core';
 import { NgbCarousel } from '@ng-bootstrap/ng-bootstrap';
+import { CmsComponentData } from '../../cms/components/cms-component-data';
 
 @Component({
   selector: 'cx-product-carousel',
@@ -24,66 +24,56 @@ import { NgbCarousel } from '@ng-bootstrap/ng-bootstrap';
   styleUrls: ['./product-carousel.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class ProductCarouselComponent extends AbstractCmsComponent
-  implements OnDestroy, OnInit {
+export class ProductCarouselComponent implements OnDestroy, OnInit {
   productGroups: Array<string[]>;
   products: { [key: string]: Observable<Product> } = {};
+  productCodes: Array<string>;
 
   resize$: Subscription;
+  dataSubscription$: Subscription;
 
   private window: Window;
 
   @ViewChild('carousel')
   carousel: NgbCarousel;
 
-  @Input()
-  productCodes: Array<string>;
-
   constructor(
-    protected cmsService: CmsService,
-    protected cd: ChangeDetectorRef,
+    public component: CmsComponentData<CmsProductCarouselComponent>,
     private productService: ProductService,
     winRef: WindowRef
   ) {
-    super(cmsService, cd);
     this.window = winRef.nativeWindow;
   }
 
   ngOnInit() {
     if (this.window) {
+      this.subscribeToData();
       this.resize$ = fromEvent(this.window, 'resize')
         .pipe(debounceTime(300))
         .subscribe(() => this.createGroups());
     }
   }
 
-  prev(): void {
+  prev() {
     this.carousel.prev();
   }
 
-  next(): void {
+  next() {
     this.carousel.next();
   }
 
-  protected fetchData(): void {
-    this.setProductCodes();
-    this.productCodes.forEach(code => {
-      this.products[code] = this.productService.get(code);
-    });
-    this.createGroups();
-    super.fetchData();
-  }
-
-  protected createGroups(): void {
+  protected createGroups() {
     const groups: Array<string[]> = [];
-    this.productCodes.forEach(product => {
-      const lastGroup: string[] = groups[groups.length - 1];
-      if (lastGroup && lastGroup.length < this.getItemsPerPage()) {
-        lastGroup.push(product);
-      } else {
-        groups.push([product]);
-      }
-    });
+    if (this.productCodes) {
+      this.productCodes.forEach(product => {
+        const lastGroup: string[] = groups[groups.length - 1];
+        if (lastGroup && lastGroup.length < this.getItemsPerPage()) {
+          lastGroup.push(product);
+        } else {
+          groups.push([product]);
+        }
+      });
+    }
     this.productGroups = groups;
   }
 
@@ -105,16 +95,22 @@ export class ProductCarouselComponent extends AbstractCmsComponent
     return itemsPerPage;
   }
 
-  protected setProductCodes(): void {
-    if (this.component && this.component.productCodes) {
-      this.productCodes = this.component.productCodes.split(' ');
-    }
+  protected subscribeToData() {
+    this.dataSubscription$ = this.component.data$.subscribe(data => {
+      this.productCodes = data.productCodes.split(' ');
+      this.productCodes.forEach(code => {
+        this.products[code] = this.productService.get(code);
+      });
+      this.createGroups();
+    });
   }
 
   ngOnDestroy() {
+    if (this.dataSubscription$) {
+      this.dataSubscription$.unsubscribe();
+    }
     if (this.resize$) {
       this.resize$.unsubscribe();
     }
-    super.ngOnDestroy();
   }
 }
