@@ -1,2 +1,2 @@
 export * from './global-message.module';
-export * from './components/global-message.component';
+export * from './global-message.component';
