@@ -6,7 +6,9 @@ import {
   RoutingService,
   UserToken,
   Title,
-  UserService
+  UserService,
+  GlobalMessageService,
+  GlobalMessageType
 } from '@spartacus/core';
 
 import { of, Observable } from 'rxjs';
@@ -64,6 +66,10 @@ class MockUserService {
   ): void {}
 }
 
+class MockGlobalMessageService {
+  remove() {}
+}
+
 describe('RegisterComponent', () => {
   let controls;
   let component: RegisterComponent;
@@ -71,6 +77,7 @@ describe('RegisterComponent', () => {
 
   let routingService: MockRoutingService;
   let userService: MockUserService;
+  let globalMessageService: MockGlobalMessageService;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
@@ -79,7 +86,8 @@ describe('RegisterComponent', () => {
       providers: [
         { provide: RoutingService, useClass: MockRoutingService },
         { provide: UserService, useClass: MockUserService },
-        { provide: AuthService, useClass: MockAuthService }
+        { provide: AuthService, useClass: MockAuthService },
+        { provide: GlobalMessageService, useClass: MockGlobalMessageService }
       ]
     }).compileComponents();
   }));
@@ -88,6 +96,7 @@ describe('RegisterComponent', () => {
     fixture = TestBed.createComponent(RegisterComponent);
     routingService = TestBed.get(RoutingService);
     userService = TestBed.get(UserService);
+    globalMessageService = TestBed.get(GlobalMessageService);
     component = fixture.componentInstance;
 
     fixture.detectChanges();
@@ -125,6 +134,14 @@ describe('RegisterComponent', () => {
           done();
         })
         .unsubscribe();
+    });
+
+    it('should remove error messages', () => {
+      spyOn(globalMessageService, 'remove').and.callThrough();
+      component.ngOnInit();
+      expect(globalMessageService.remove).toHaveBeenCalledWith(
+        GlobalMessageType.MSG_TYPE_ERROR
+      );
     });
 
     it('should go to redirect url after registration', () => {
